CanAgent/requirements.txt
"""
Requirements file for Notion Agent System
"""

requests>=2.25.0
beautifulsoup4>=4.9.3
python-dotenv>=0.15.0
